+++ src/components/bestProducts/BestProductCard.tsx
@@ -1,5 +1,5 @@
 import CardBest from "./CardBest";
-import DefaultImg from "../../core/assets/images/img_default.png";
+import DefaultImg from "../../core/assets/images/product/img_default.png";
 import { ProductItem } from "core/Interface/Product";
 
 interface BestProductCardProps {
